Let PlaybackController handle listening for service start
