+++ app/src/main/java/de/danoeh/antennapod/activity/MediaplayerActivity.java
@@ -38,7 +38,6 @@ import com.joanzapata.iconify.fonts.FontAwesomeIcons;
 import java.util.Locale;
 
 import de.danoeh.antennapod.R;
-import de.danoeh.antennapod.core.event.ServiceEvent;
 import de.danoeh.antennapod.core.feed.FeedItem;
 import de.danoeh.antennapod.core.feed.FeedMedia;
 import de.danoeh.antennapod.core.feed.MediaType;
@@ -68,9 +67,6 @@ import io.reactivex.Observable;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.Disposable;
 import io.reactivex.schedulers.Schedulers;
-import org.greenrobot.eventbus.EventBus;
-import org.greenrobot.eventbus.Subscribe;
-import org.greenrobot.eventbus.ThreadMode;
 
 
 /**
@@ -288,7 +284,6 @@ public abstract class MediaplayerActivity extends CastEnabledActivity implements
         controller.init();
         loadMediaInfo();
         onPositionObserverUpdate();
-        EventBus.getDefault().register(this);
     }
 
     @Override
@@ -301,7 +296,6 @@ public abstract class MediaplayerActivity extends CastEnabledActivity implements
         if (disposable != null) {
             disposable.dispose();
         }
-        EventBus.getDefault().unregister(this);
         super.onStop();
     }
 
@@ -667,16 +661,6 @@ public abstract class MediaplayerActivity extends CastEnabledActivity implements
         super.onResume();
         Log.d(TAG, "onResume()");
         StorageUtils.checkStorageAvailability(this);
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onEventMainThread(ServiceEvent event) {
-        Log.d(TAG, "onEvent(" + event + ")");
-        if (event.action == ServiceEvent.Action.SERVICE_STARTED) {
-            if (controller != null) {
-                controller.init();
-            }
-        }
     }
 
     /**
